Fix: change the app folder structure
