--- server/controllers/controller.js
@@ -1,5 +1,0 @@
-class Controller {
-
-}
-
-export default Controller;
