+++ server/controllers/mealController/getMeals.js
@@ -1,0 +1,10 @@
+import Meals from '../../dummy-models/meals';
+
+const getMeals = (req, res) => res.status(200).json({
+  status: 'success',
+  data: {
+    meals: Meals,
+  },
+});
+
+export default getMeals;
