+++ server/controllers/mealController/index.js
@@ -1,0 +1,15 @@
+import createMeal from './createMeal';
+import getMeals from './getMeals';
+import getMeal from './getMeal';
+import updateMeal from './updateMeal';
+import deleteMeal from './deleteMeal';
+
+class MealController {}
+
+MealController.createMeal = createMeal;
+MealController.getMeals = getMeals;
+MealController.getMeal = getMeal;
+MealController.updateMeal = updateMeal;
+MealController.deleteMeal = deleteMeal;
+
+export default MealController;
