+++ server/controllers/menuController/index.js
@@ -1,0 +1,13 @@
+import createMenu from './createMenu';
+import getMenus from './getMenus';
+import getTodaysMenu from './getTodaysMenu';
+import getSpecificDayMenu from './getSpecificDayMenu';
+
+class MenuController {}
+
+MenuController.createMenu = createMenu;
+MenuController.getMenus = getMenus;
+MenuController.getTodaysMenu = getTodaysMenu;
+MenuController.getSpecificDayMenu = getSpecificDayMenu;
+
+export default MenuController;
