--- server/routes/index.js
@@ -1,9 +1,0 @@
-import MealsRoutes from './mealsRoutes';
-import MenuRoutes from './menuRoutes';
-import MenusRoutes from './menuRoutes';
-
-export {
-	MealsRoutes,
-	MenusRoutes,
-	MenuRoutes,
-}
