+++ server/routes/menus.js
@@ -5,9 +5,7 @@ const router = express.Router();
 
 // These routes are to be prepended with "/menus"
 router.post('/', MenuController.createMenu);
-
 router.get('/', MenuController.getMenus);
-
 router.get('/:date', MenuController.getSpecificDayMenu);
 
 // router.put('/:menuId', () => {});
