+++ server/tests/controllers/MealController/testCreateMeal.js
@@ -4,7 +4,6 @@ import sinonChai from 'sinon-chai';
 import { mockReq, mockRes } from 'sinon-express-mock';
 
 import Meals from '../../../dummy-models/meals';
-import Controller from '../../../controllers/controller';
 import MealController from '../../../controllers/mealController';
