+++ server/tests/controllers/MealController/testGetMeal.js
@@ -3,7 +3,6 @@ import sinon from 'sinon';
 import sinonChai from 'sinon-chai';
 import { mockReq, mockRes } from 'sinon-express-mock';
 
-import Controller from '../../../controllers/controller';
 import MealController from '../../../controllers/mealController';
 
 
@@ -26,6 +25,7 @@ const badRequest = {
 
 
 const req = mockReq(request);
+const badReq = mockReq(badRequest);
 const res = mockRes();
 
 describe('getMeals method', () => {
