--- server/tests/controllers/testController.js
@@ -1,10 +1,0 @@
-import chai from 'chai';
-import Controller from '../../controllers/controller';
-
-chai.should();
-
-describe('Base controller', () => {
-  it('should be a function', () => {
-    Controller.should.be.a('function');
-  });
-});
